OneStage/yolo/deep_sort_yolov3/requirements.txt
Keras==2.3.1
tensorflow-gpu==1.15.2
numpy==1.16.0
opencv-python==3.4.4.19
scikit-learn==0.21.2
scipy==1.1.0
Pillow
torch==1.3.0
torchvision==0.4.1
